+++ wow_world_base/src/inner/wrath/mod.rs
@@ -26,6 +26,7 @@ pub use crate::shared::arena_type_tbc_wrath::*;
 pub use crate::shared::auction_command_action_vanilla_tbc_wrath::*;
 pub use crate::shared::auction_command_result_vanilla_tbc_wrath::*;
 pub use crate::shared::auction_command_result_two_vanilla_tbc_wrath::*;
+pub use crate::shared::auction_house_vanilla_tbc_wrath::*;
 pub(crate) mod aura_mod;
 pub use aura_mod::*;
 pub(crate) mod aura_type;