+++ wow_world_base/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+* `AuctionHouse` type for Vanilla, TBC and Wrath.
 
 ## [0.1.1] - 2023-05-20